+++ src/full/Agda/TypeChecking/Serialise/Instances/Common.hs
@@ -173,16 +173,25 @@ instance EmbPrj FileType where
     [3] -> valuN OrgFileType
     _   -> malformed
 
-instance EmbPrj Language where
-  icod_ WithoutK = icodeN'  WithoutK
-  icod_ WithK    = icodeN 0 WithK
-  icod_ Cubical  = icodeN 1 Cubical
+instance EmbPrj Cubical where
+  icod_ CErased = icodeN'  CErased
+  icod_ CFull   = icodeN 0 CFull
 
   value = vcase $ \case
-    []  -> valuN WithoutK
-    [0] -> valuN WithK
-    [1] -> valuN Cubical
+    []  -> valuN CErased
+    [0] -> valuN CFull
     _   -> malformed
+
+instance EmbPrj Language where
+  icod_ WithoutK    = icodeN'  WithoutK
+  icod_ WithK       = icodeN 0 WithK
+  icod_ (Cubical a) = icodeN 1 Cubical a
+
+  value = vcase $ \case
+    []     -> valuN WithoutK
+    [0]    -> valuN WithK
+    [1, a] -> valuN Cubical a
+    _      -> malformed
 
 instance EmbPrj AbsolutePath where
   icod_ file = do